LayoutTests/storage/indexeddb/modern/createobjectstore-basic.html

This test starts some version change transactions, creates some object stores, and variably commits or aborts the version change transactions.<br>
At various stages it verifies the object stores in the database are as-expected.<br>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

function dumpObjectStores(database) {
    var list = database.objectStoreNames;
    alert("Object store names:");
    for (var i = 0; i < list.length; ++i) { 
        alert(list[i]);
    }
}

var createRequest = window.indexedDB.open("CreateObjectStoreTestDatabase", 1);

createRequest.onupgradeneeded = function(event) {
    alert("Initial upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    var versionTransaction = createRequest.transaction;
    var database = event.target.result;
    var objectStore = database.createObjectStore("FirstAbortedObjectStore");
    var request = objectStore.put("foo", "bar");

    request.onsuccess = function(event) {
        alert("Put succeeded");
        versionTransaction.abort();
    }
    request.onerror = function(event) {
        alert("Put failed - " + event);
        done();
    }
    
    dumpObjectStores(database);    
    
    versionTransaction.onabort = function(event) {
        alert("Initial upgrade versionchange transaction aborted");
        dumpObjectStores(database);
        continueTest1();
        database.close();
    }

    versionTransaction.oncomplete = function(event) {
        alert("Initial upgrade versionchange transaction unexpected complete");
        done();
    }

    versionTransaction.onerror = function(event) {
        alert("Initial upgrade versionchange transaction unexpected error" + event);
        done();
    }
}

function continueTest1()
{
    createRequest = window.indexedDB.open("CreateObjectStoreTestDatabase", 1);

    createRequest.onupgradeneeded = function(event) {
        alert("Second upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

        var versionTransaction = createRequest.transaction;
        var database = event.target.result;
        dumpObjectStores(database);
        var objectStore = database.createObjectStore("FirstCommittedObjectStore");

        versionTransaction.onabort = function(event) {
            alert("Second upgrade versionchange transaction unexpected abort");
            done();
        }

        versionTransaction.oncomplete = function(event) {
            alert("Second upgrade versionchange transaction complete");
            dumpObjectStores(database);
            continueTest2();
            database.close();
        }

        versionTransaction.onerror = function(event) {
            alert("Second upgrade versionchange transaction unexpected error" + event);
            done();
        }
    }
}

function continueTest2()
{
    createRequest = window.indexedDB.open("CreateObjectStoreTestDatabase", 2);

    createRequest.onupgradeneeded = function(event) {
        alert("Third upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

        var versionTransaction = createRequest.transaction;
        var database = event.target.result;
        var objectStore = database.createObjectStore("SecondCommittedObjectStore");

        dumpObjectStores(database);    
    
        versionTransaction.onabort = function(event) {
            alert("Third upgrade versionchange transaction unexpected abort");
            done();
        }

        versionTransaction.oncomplete = function(event) {
            alert("Third upgrade versionchange transaction complete");
            dumpObjectStores(database);
            database.close(); 
            continueTest3();
        }

        versionTransaction.onerror = function(event) {
            alert("Third upgrade versionchange transaction unexpected error" + event);
            done();
        }
    }
}

function continueTest3()
{
    createRequest = window.indexedDB.open("CreateObjectStoreTestDatabase", 3);

    createRequest.onupgradeneeded = function(event) {
        alert("Fourth upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);
        var database = event.target.result;
        dumpObjectStores(database);
        done();
    }
}

</script>
