LayoutTests/storage/indexeddb/modern/createobjectstore-basic-expected.txt

ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Object store names:
ALERT: FirstAbortedObjectStore
ALERT: Put succeeded
ALERT: Initial upgrade versionchange transaction aborted
ALERT: Object store names:
ALERT: Second upgrade needed: Old version - 0 New version - 1
ALERT: Object store names:
ALERT: Second upgrade versionchange transaction complete
ALERT: Object store names:
ALERT: FirstCommittedObjectStore
ALERT: Third upgrade needed: Old version - 1 New version - 2
ALERT: Object store names:
ALERT: FirstCommittedObjectStore
ALERT: SecondCommittedObjectStore
ALERT: Third upgrade versionchange transaction complete
ALERT: Object store names:
ALERT: FirstCommittedObjectStore
ALERT: SecondCommittedObjectStore
ALERT: Fourth upgrade needed: Old version - 2 New version - 3
ALERT: Object store names:
ALERT: FirstCommittedObjectStore
ALERT: SecondCommittedObjectStore
ALERT: Done
This test starts some version change transactions, creates some object stores, and variably commits or aborts the version change transactions.
At various stages it verifies the object stores in the database are as-expected.
